requirements.txt
pywin32 # for windows
Pillow
replicate
python-dotenv
certifi
pyserial
argparse
pythoncom